Remove dependency on can-map and can-stache
