+++ package.json
@@ -60,9 +60,7 @@
     "can-compute": "^3.0.0-pre.2",
     "can-construct": "^3.0.0-pre.2",
     "can-control": "^3.0.0-pre.1",
-    "can-map": "^3.0.0-pre.2",
     "can-observe-info": "^3.0.0-pre.4",
-    "can-stache": "^3.0.0-pre.1",
     "can-stache-bindings": "^3.0.0-pre.1",
     "can-util": "^3.0.0-pre.7",
     "can-view-callbacks": "^3.0.0-pre.1",
@@ -71,6 +69,8 @@
     "can-view-scope": "^3.0.0-pre.1"
   },
   "devDependencies": {
+  	"can-map": "^3.0.0-pre.2",
+  	"can-stache": "^3.0.0-pre.1",
     "can-define": "^0.3.0",
     "can-event": "^3.0.0-pre.1",
     "can-list": "^3.0.0-pre.1",
